% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci.R
\name{bootci}
\alias{bootci}
\title{Compute Confidence Intervalls via Bootstrap}
\usage{
bootci(x, k, n)
}
\arguments{
\item{x}{Numeric integer. The sample.}

\item{k}{Integer. Number of samples to compute mean (for each iteration).}

\item{n}{Integer. Number of iterations.}
}
\description{
Compute Confidence Intervalls via Bootstrap
}
\examples{
bootci(iris$Sepal.Length, 50, 10000)
}
