% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_m.R
\name{bootci_m}
\alias{bootci_m}
\title{Compute Confidence Intervalls via Bootstrap for multiple fractions}
\usage{
bootci_m(x, iteration = 1000, fraction = c(0.05, 0.25, 0.5), digits = 2)
}
\arguments{
\item{x}{Numeric Vector.}

\item{iteration}{Integer. Specify number of iterations.}

\item{fraction}{Numeric vector of fractions. e.g. 0.5 means that ~50 percent of original sample size will be used to compute mean value for each iteration.}

\item{digits}{Integer. Specify number of digits.}
}
\description{
Choose multiple fractions of sample size to track conversion of CI
}
\note{
For faster computation of a single fraction use bootci instead.
}
\references{
Crawley, Michael J. The R Book. Chichester, England; Hoboken, N.J.: Wiley, 2007.
}
