% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(value)
}
\arguments{
\item{value}{Numeric vector}
}
\description{
Kurtosis
}
\examples{
kurtosis(iris$Sepal.Length)
}
\references{
Crawley, Michael J. The R Book. Chichester, England; Hoboken, N.J.: Wiley, 2007.
}
