% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Skewness}
\usage{
skew(value)
}
\arguments{
\item{value}{Numeric vector.}
}
\description{
Skewness
}
\examples{
skew(iris$Sepal.Length)
}
\references{
https://stat.ethz.ch/pipermail/r-help/1999-July/004529.html
}
